requirements.txt
numpy
matplotlib
ArgumentStack
piecash
pyyaml 
requests 
pandas